// iped-webui/src/app/core/api/generated/model/exportJobRequest.ts
/**
 * IPED Web API - Phase 1 Initial Contracts
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ExportJobRequest { 
    caseId: string;
    itemIds?: Array<string>;
    format?: ExportJobRequestFormatEnum;
}
export enum ExportJobRequestFormatEnum {
    zip = 'zip',
    csv = 'csv',
    report = 'report'
};