// iped-webui/src/app/core/api/generated/model/jobAcceptedResponse.ts
/**
 * IPED Web API - Phase 1 Initial Contracts
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface JobAcceptedResponse { 
    jobId: string;
    status: JobAcceptedResponseStatusEnum;
}
export enum JobAcceptedResponseStatusEnum {
    accepted = 'accepted',
    running = 'running'
};