// iped-webui/src/app/core/api/generated/model/jobStatus.ts
/**
 * IPED Web API - Phase 1 Initial Contracts
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface JobStatus { 
    jobId: string;
    status: JobStatusStatusEnum;
    progress?: number;
    message?: string;
    outputUrl?: string;
}
export enum JobStatusStatusEnum {
    accepted = 'accepted',
    running = 'running',
    completed = 'completed',
    failed = 'failed',
    cancelled = 'cancelled'
};